refactor: add save computer in computer repository
